- Refactored EVSProto implementation out from evs.cpp - Added evs files in makefiles
git-svn-id: https://devel.codership.com/svn/galera/trunk@604 9148e567-b586-4176-b9b8-4e36cb4ec61b

+++ galeracomm/vs/src/evs_message.hpp
@@ -448,7 +448,7 @@ public:
 	    if ((off = source_view.read(buf, buflen, off)) == 0)
 		return 0;
 	    if (type == JOIN || type == INSTALL) {
-		size_t n;
+		uint32_t n;
 		if ((off = read_uint32(buf, buflen, off, &n)) == 0)
 		    return 0;
 		delete instances;
